Add token authentication to Active Storage controllers [SCI-5123]
