+++ app/controllers/active_storage/custom_base_controller.rb
@@ -3,6 +3,10 @@
 # The base controller for all ActiveStorage controllers.
 module ActiveStorage
   class CustomBaseController < ApplicationController
+    include TokenAuthentication
     include ActiveStorage::SetCurrent
+
+    prepend_before_action :authenticate_request!, if: -> { request.headers['Authorization'].present? }
+    skip_before_action :authenticate_user!, if: -> { current_user.present? }
   end
 end
