+++ app/controllers/api/api_controller.rb
@@ -2,8 +2,8 @@
 
 module Api
   class ApiController < ActionController::API
-    attr_reader :iss
-    attr_reader :token
+    include TokenAuthentication
+
     attr_reader :current_user
 
     before_action :authenticate_request!, except: %i(status health)
@@ -52,46 +52,6 @@ module Api
         response[:versions] << { version: ver, baseUrl: "/api/#{ver}/" }
       end
       render json: response, status: :ok
-    end
-
-    private
-
-    def azure_jwt_auth
-      return unless iss =~ %r{windows.net/|microsoftonline.com/}
-      token_payload, = Api::AzureJwt.decode(token)
-      @current_user = User.from_azure_jwt_token(token_payload)
-      unless current_user
-        raise JWT::InvalidPayload, I18n.t('api.core.no_azure_user_mapping')
-      end
-    end
-
-    def authenticate_request!
-      @token = request.headers['Authorization']&.sub('Bearer ', '')
-      unless @token
-        raise JWT::VerificationError, I18n.t('api.core.missing_token')
-      end
-
-      @iss = CoreJwt.read_iss(token)
-      raise JWT::InvalidPayload, I18n.t('api.core.no_iss') unless @iss
-
-      Extends::API_PLUGABLE_AUTH_METHODS.each do |auth_method|
-        method(auth_method).call
-        return true if current_user
-      end
-
-      # Default token implementation
-      unless iss == Rails.configuration.x.core_api_token_iss
-        raise JWT::InvalidPayload, I18n.t('api.core.wrong_iss')
-      end
-      payload = CoreJwt.decode(token)
-      @current_user = User.find_by_id(payload['sub'])
-      unless current_user
-        raise JWT::InvalidPayload, I18n.t('api.core.no_user_mapping')
-      end
-    end
-
-    def auth_params
-      params.permit(:grant_type, :email, :password)
     end
   end
 end
